Remove Middleware and Postware
