+++ app/src/main/java/com/github/skyfe79/android/library/app/examples/counter/CounterViewModel.kt
@@ -13,9 +13,7 @@ class CounterViewModel: RootViewModelType<CounterState>() {
     override fun setupStore() {
         store.set(
             initialState = CounterState(0),
-            middlewares = arrayOf(),
-            reducers = arrayOf(::countReducer),
-            postwares = arrayOf()
+            reducers = arrayOf(::countReducer)
         )
     }
