+++ reactcomponentkit/src/main/java/com/github/skyfe79/android/reactcomponentkit/redux/Middleware.kt
@@ -1,3 +1,3 @@
 package com.github.skyfe79.android.reactcomponentkit.redux
 
-typealias Middleware<STATE> = (STATE, Action) -> STATE
+//typealias Middleware<STATE> = (STATE, Action) -> STATE
