+++ reactcomponentkit/src/main/java/com/github/skyfe79/android/reactcomponentkit/redux/Postware.kt
@@ -1,3 +1,3 @@
 package com.github.skyfe79.android.reactcomponentkit.redux
 
-typealias Postware<STATE> = (STATE, Action) -> STATE
+//typealias Postware<STATE> = (STATE, Action) -> STATE
